revert initcpiocfg changes
This got once again merged from the calamares team without any testing.
After reporting already a bunch of issues with it, there is still a issue that it breaks encrypted installation.
The PR got merged, because someone requested it for encrypted installation with LVM.

It really does not fit in my head, how something gets merged and the reason to create this feature was not tested.

Signed-off-by: Peter Jung <admin@ptr1337.dev>