[NEEDS TESTING] Change Btrfs subvolume layout
The idea here is simple: Create Btrfs subvolumes in the same path where they are mounted to. This way, having them unmounted (as is the case when, for example, chrooting into the system for system repair) will not create hidden junk files in the paths where the subvolumes are normally mounted to.
The purpose of the subvolumes remains completely fulfilled, namely preventing the subvolume mount paths to be part of system snapshots.
@home subvolume is excluded from this change, so it remains fully compatible with Timeshift.
I have used this setup myself for a long time, but I'm not sure if this patch will actually work as intended, so probably needs testing.